fix to cluster timing, down to ~0.05 sec/event from 0.12, can do better by refactorizing PositionCalc
